refactor(SVDP): rename npu solution
